+++ sample_apps/simple_app/example_app.mak
@@ -67,7 +67,6 @@ endif
 
 # Note:  You can omit all the files below in d_obj except $(O)/example_app.o,
 #  if you link in the certifier library certifier.a.
-ifdef NEW_API
 dobj = $(O)/example_app.o $(O)/certifier.pb.o $(O)/certifier.o $(O)/certifier_proofs.o \
        $(O)/support.o $(O)/simulated_enclave.o $(O)/application_enclave.o $(O)/cc_helpers.o \
        $(O)/cc_useful.o
@@ -75,15 +74,6 @@ dobj = $(O)/example_app.o $(O)/certifier.pb.o $(O)/certifier.o $(O)/certifier_pr
 robj = $(O)/example_key_rotation.o $(O)/certifier.pb.o $(O)/certifier.o $(O)/certifier_proofs.o \
        $(O)/support.o $(O)/simulated_enclave.o $(O)/application_enclave.o $(O)/cc_helpers.o \
        $(O)/cc_useful.o
-else
-dobj = $(O)/example_app_old_api.o $(O)/certifier.pb.o $(O)/certifier.o $(O)/certifier_proofs.o \
-       $(O)/support.o $(O)/simulated_enclave.o $(O)/application_enclave.o $(O)/cc_helpers.o \
-       $(O)/cc_useful.o
-
-robj = $(O)/example_key_rotation_old_api.o $(O)/certifier.pb.o $(O)/certifier.o $(O)/certifier_proofs.o \
-       $(O)/support.o $(O)/simulated_enclave.o $(O)/application_enclave.o $(O)/cc_helpers.o \
-       $(O)/cc_useful.o
-endif
 
 all:	example_app.exe example_key_rotation.exe
 clean:
@@ -111,7 +101,6 @@ $(O)/certifier.pb.o: $(US)/certifier.pb.cc $(I)/certifier.pb.h
 	@echo "\ncompiling $<"
 	$(CC) $(CFLAGS_NOERROR) -o $(@D)/$@ -c $<
 
-ifdef NEW_API
 $(O)/example_app.o: $(COMMON_SRC)/example_app.cc $(I)/certifier.h $(US)/certifier.pb.cc
 	@echo "\ncompiling $<"
 	$(CC) $(CFLAGS) -o $(@D)/$@ -c $<
@@ -119,15 +108,6 @@ $(O)/example_app.o: $(COMMON_SRC)/example_app.cc $(I)/certifier.h $(US)/certifie
 $(O)/example_key_rotation.o: $(US)/example_key_rotation.cc $(I)/certifier.h $(US)/certifier.pb.cc
 	@echo "\ncompiling $<"
 	$(CC) $(CFLAGS) -o $(@D)/$@ -c $<
-else
-$(O)/example_app_old_api.o: $(COMMON_SRC)/example_app_old_api.cc $(I)/certifier.h $(US)/certifier.pb.cc
-	@echo "\ncompiling $<"
-	$(CC) $(CFLAGS) -o $(@D)/$@ -c $<
-
-$(O)/example_key_rotation_old_api.o: $(US)/example_key_rotation_old_api.cc $(I)/certifier.h $(US)/certifier.pb.cc
-	@echo "\ncompiling $<"
-	$(CC) $(CFLAGS) -o $(@D)/$@ -c $<
-endif
 
 $(O)/certifier.o: $(S)/certifier.cc $(I)/certifier.pb.h $(I)/certifier.h
 	@echo "\ncompiling $<"
